feat: accept yup validate options
This way, the caller can pass in any `validate` function options to Yup.

For example, I can pass in `abortEarly: false` so that all validation
errors would be returned instead of only the first one. It's worth
noting that in that case, I also need to use a custom `formatError`
function to correctly return the nested errors.